+++ cfgov/ask_cfpb/models/__init__.py
@@ -8,3 +8,4 @@ from ask_cfpb.models.pages import (
     get_reusable_text_snippet, get_standard_text, validate_page_number
 )
 from ask_cfpb.models.search import AskSearch
+from ask_cfpb.models.snippets import GlossaryTerm
