+++ cfgov/v1/wagtail_hooks.py
@@ -20,13 +20,12 @@ from wagtail.wagtailcore.whitelist import attribute_rule
 
 from scripts import export_enforcement_actions
 
+from ask_cfpb.models.snippets import GlossaryTerm
 from v1.admin_views import ExportFeedbackView, manage_cdn
 from v1.models.menu_item import MenuItem as MegaMenuItem
 from v1.models.portal_topics import PortalCategory, PortalTopic
 from v1.models.resources import Resource
-from v1.models.snippets import (
-    Contact, GlossaryTerm, RelatedResource, ReusableText
-)
+from v1.models.snippets import Contact, RelatedResource, ReusableText
 from v1.util import util
